requirements.txt
streamlit==1.21.0
Pillow==9.2.0
numpy==1.22.3
opencv-python==4.9.0.80
tensorflow==2.12.0